+++ tests/integration/pre_run_script.sh
@@ -6,7 +6,7 @@
 # Pre-run script for integration test operator-workflows action.
 # https://github.com/canonical/operator-workflows/blob/main/.github/workflows/integration_test.yaml
 
-# OpenSearch charms are deployed on lxd and Wazuh Server charm is deployed on microk8s.
+# OpenSearch charms are deployed on lxd and Wazuh Server charm is deployed on Canonical K8S.
 
 set -euo pipefail
 
@@ -14,19 +14,10 @@ TESTING_MODEL="$(juju switch)"
 
 # lxd should be install and init by a previous step in integration test action.
 echo "bootstrapping lxd juju controller"
-# Change microk8s default file limits
-echo "ulimit -n 458752" | sudo tee -a /var/snap/microk8s/current/args/containerd-env
-sudo snap restart microk8s
-sg snap_microk8s -c "microk8s status --wait-ready"
-sg snap_microk8s -c "juju bootstrap localhost localhost"
+juju bootstrap localhost localhost
 
 echo "Switching to testing model"
-sg snap_microk8s -c "juju switch $TESTING_MODEL"
-
-
-IPADDR=$(ip -4 -j route get 2.2.2.2 | jq -r '.[] | .prefsrc')
-sudo microk8s enable "metallb:$IPADDR-$IPADDR"
-sudo microk8s status
+juju switch "$TESTING_MODEL"
 
 # https://charmhub.io/opensearch/docs/t-set-up#set-parameters-on-the-host-machine
 sudo tee -a /etc/sysctl.conf > /dev/null <<EOT