+++ docs/changelog.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 Each revision is versioned by the date of the revision.
 
+## 2025-06-19
+
+### Updated
+
+- Modified the CI and `pre_run_script.sh` to use Canonical K8s instead of Microk8s.
+
 ## 2025-06-16
 
 ### Updated